Factor out TimeSeries and Quantiles and prepare to test those with specific tests, so that Metrics simply tests observance and encoding
